restructuring index.ts
